refactor: riot.api.ts 파일 삭제 및 import 경로 수정
